cloud_run_test/usd_krw_docker/requirements.txt
functions-framework==3.*
pandas
psycopg2
sqlalchemy
pykrx
plotly

flask

bs4
pandas-gbq
google-cloud-bigquery
google-cloud-storage
